Move codes from RecordListAdapter to MainActivity
